+++ riscv-rt/tests/ui/core_interrupt/pass_core_interrupt.rs
@@ -1,0 +1,11 @@
+use riscv::interrupt::Interrupt::*;
+
+#[riscv_rt::core_interrupt(SupervisorSoft)]
+fn simple_interrupt() {}
+
+#[riscv_rt::core_interrupt(SupervisorTimer)]
+unsafe fn no_return_interrupt() -> ! {
+    loop {}
+}
+
+fn main() {}